Move trainer and combiner schemas over. (still need to fix tests, maybe some imports).
