+++ ludwig/schema/combiners/__init__.py
@@ -1,0 +1,8 @@
+# Pull configs here for convenient imports in other modules:
+from comparator import ComparatorCombinerConfig  # noqa: F401
+from concat import ConcatCombinerConfig  # noqa: F401
+from sequence import SequenceCombinerConfig  # noqa: F401
+from sequence_concat import SequenceConcatCombinerConfig  # noqa: F401
+from tab_transformer import TabTransformerCombinerConfig  # noqa: F401
+from tabnet import TabNetCombinerConfig  # noqa: F401
+from transformer import TransformerCombinerConfig  # noqa: F401
