+++ ludwig/schema/combiners/base.py
@@ -1,0 +1,7 @@
+from ludwig.schema import utils as schema_utils
+
+
+class BaseCombinerConfig(schema_utils.BaseMarshmallowConfig):
+    """Base combiner config class."""
+
+    pass
